refactor: auto-split test_terrain.py
