+++ tests/unit/terrain/test_terrain_region.py
@@ -1,0 +1,62 @@
+"""Unit tests for terrain modeling system (TDD: Tests First).
+
+Following the Pragmatic Programmer principles:
+- DRY (Don't Repeat Yourself)
+- Design by Contract
+- Orthogonality
+- Test-Driven Development
+
+These tests define the expected behavior BEFORE implementation.
+"""
+
+from __future__ import annotations
+
+import json
+import math
+from pathlib import Path
+
+import numpy as np
+import pytest
+
+# Tests written first - imports will fail until implementation exists
+from src.shared.python.physics.terrain import (
+    ElevationMap,
+    SurfaceMaterial,
+    Terrain,
+    TerrainConfig,
+    TerrainPatch,
+    TerrainRegion,
+    TerrainType,
+    create_flat_terrain,
+    create_sloped_terrain,
+    create_terrain_from_config,
+)
+
+
+class TestTerrainRegion:
+    """Test terrain regions with complex shapes."""
+
+    def test_circular_region(self) -> None:
+        """Create a circular terrain region (e.g., green)."""
+        region = TerrainRegion.circle(
+            terrain_type=TerrainType.GREEN,
+            center_x=100.0,
+            center_y=100.0,
+            radius=15.0,
+        )
+
+        assert region.contains(100.0, 100.0)  # Center
+        assert region.contains(110.0, 100.0)  # On edge
+        assert not region.contains(120.0, 100.0)  # Outside
+
+    def test_polygon_region(self) -> None:
+        """Create a polygon-shaped terrain region."""
+        # Triangle
+        vertices = [(0.0, 0.0), (10.0, 0.0), (5.0, 10.0)]
+        region = TerrainRegion.polygon(
+            terrain_type=TerrainType.BUNKER,
+            vertices=vertices,
+        )
+
+        assert region.contains(5.0, 3.0)  # Inside triangle
+        assert not region.contains(0.0, 10.0)  # Outside
